+++ src/interpreter/map-validator.ts
@@ -23,6 +23,7 @@ import {
   PrimitiveLiteralNode,
   SetStatementNode,
 } from '@superfaceai/ast';
+import createDebug from 'debug';
 import * as ts from 'typescript';
 
 import { RETURN_CONSTRUCTS } from './constructs';
@@ -49,6 +50,8 @@ import {
   mergeVariables,
   validateObjectStructure,
 } from './utils';
+
+const debug = createDebug('superface-parser:map-validator');
 
 function assertUnreachable(node: never): never;
 function assertUnreachable(node: MapASTNode): never {
@@ -96,6 +99,8 @@ export class MapValidator implements MapAstVisitor {
   visit(node: LiteralNode | AssignmentNode): boolean;
   visit(node: MapASTNode): void;
   visit(node: MapASTNode): boolean | void {
+    debug('Visiting node: ' + node.kind);
+
     switch (node.kind) {
       case 'MapDocument':
         return this.visitMapDocumentNode(node);
@@ -194,7 +199,7 @@ export class MapValidator implements MapAstVisitor {
   }
 
   visitMapHeaderNode(node: MapHeaderNode): void {
-    const { scope, name, version } = this.profileOutput.header;
+    const { scope, name, version: profileVersion } = this.profileOutput.header;
 
     if (
       (scope && scope !== node.profile.scope) ||
@@ -222,14 +227,17 @@ export class MapValidator implements MapAstVisitor {
     }
 
     // map should be compatible with every patch version of a profile, therefore it should ignore patch version
-    const { major, minor } = node.profile.version;
-    if (major !== version.major || minor !== version.minor) {
+    const mapVersion = node.profile.version;
+    if (
+      mapVersion.major !== profileVersion.major ||
+      mapVersion.minor !== profileVersion.minor
+    ) {
       this.errors.push({
         kind: 'wrongProfileVersion',
         context: {
           path: this.getPath(node),
-          expected: version,
-          actual: node.profile.version,
+          expected: profileVersion,
+          actual: mapVersion,
         },
       });
     }