+++ src/interpreter/utils.ts
@@ -33,7 +33,8 @@ import {
 
 export function composeVersion(version: VersionStructure): string {
   return (
-    `${version.major}.${version.minor}.${version.patch}` +
+    `${version.major}.${version.minor}` +
+    (version.patch !== undefined ? `.${version.patch}` : '') +
     (version.label ? `-${version.label}` : '')
   );
 }
@@ -64,9 +65,9 @@ export function formatIssues(issues?: ValidationIssue[]): string {
           return `${location} - Wrong Profile Name: expected ${issue.context.expected}, but got ${issue.context.actual}`;
 
         case 'wrongProfileVersion':
-          return `${location} - Wrong Profile Version: expected ${composeVersion(
+          return `${location} - Wrong Profile Version: profile is ${composeVersion(
             issue.context.expected
-          )}, but got ${composeVersion(issue.context.actual)}`;
+          )}, but map requests ${composeVersion(issue.context.actual)}`;
 
         case 'mapNotFound':
           return `${location} - Map not found: ${issue.context.expected}`;